issue #139 feat: move reusable getFirstDateOfWeek in utils
